Rename sparkUtils into spark_utils
