+++ bin/access_science_db.py
@@ -20,7 +20,7 @@ import argparse
 import json
 
 from fink_broker.parser import getargs
-from fink_broker.sparkUtils import init_sparksession
+from fink_broker.spark_utils import init_sparksession
 from fink_broker.logging_utils import get_fink_logger, inspect_application
