+++ bin/daily_stats.py
@@ -22,7 +22,7 @@ import pandas as pd
 
 import pyspark.sql.functions as F
 
-from fink_broker.sparkUtils import init_sparksession
+from fink_broker.spark_utils import init_sparksession
 from fink_broker.hbase_utils import push_to_hbase
 from fink_broker.parser import getargs
 from fink_broker.logging_utils import get_fink_logger, inspect_application
