+++ bin/distribute_elasticc.py
@@ -30,7 +30,7 @@ from fink_broker import __version__ as fbvsn
 from fink_science import __version__ as fsvsn
 
 from fink_broker.parser import getargs
-from fink_broker.sparkUtils import init_sparksession, connect_to_raw_database
+from fink_broker.spark_utils import init_sparksession, connect_to_raw_database
 from fink_broker.distribution_utils import get_kafka_df
 from fink_broker.logging_utils import get_fink_logger, inspect_application
 from fink_broker.partitioning import convert_to_millitime
