+++ bin/download_tns.py
@@ -20,7 +20,7 @@ import argparse
 import os
 
 from fink_tns.utils import download_catalog
-from fink_broker.sparkUtils import init_sparksession
+from fink_broker.spark_utils import init_sparksession
 from fink_broker.hbase_utils import add_row_key, push_to_hbase
 from fink_broker.parser import getargs
