+++ bin/index_archival.py
@@ -39,7 +39,7 @@ from fink_broker.hbase_utils import load_ztf_index_cols
 from fink_broker.hbase_utils import load_ztf_crossmatch_cols
 from fink_broker.hbase_utils import select_relevant_columns
 from fink_broker.hbase_utils import bring_to_current_schema
-from fink_broker.sparkUtils import init_sparksession, load_parquet_files
+from fink_broker.spark_utils import init_sparksession, load_parquet_files
 from fink_broker.logging_utils import get_fink_logger, inspect_application
 
 from fink_filters.classification import extract_fink_classification
