+++ bin/index_sso_resolver.py
@@ -22,7 +22,7 @@ import datetime
 import pandas as pd
 import numpy as np
 
-from fink_broker.sparkUtils import init_sparksession
+from fink_broker.spark_utils import init_sparksession
 from fink_broker.hbase_utils import push_to_hbase
 from fink_broker.logging_utils import get_fink_logger, inspect_application
