+++ bin/push_to_tns.py
@@ -20,7 +20,7 @@ import requests
 import os
 
 from fink_broker.parser import getargs
-from fink_broker.sparkUtils import init_sparksession, load_parquet_files
+from fink_broker.spark_utils import init_sparksession, load_parquet_files
 from fink_broker.logging_utils import get_fink_logger, inspect_application
 
 from fink_filters.classification import extract_fink_classification
