+++ bin/raw2science.py
@@ -30,8 +30,8 @@ import time
 
 from fink_broker import __version__ as fbvsn
 from fink_broker.parser import getargs
-from fink_broker.sparkUtils import init_sparksession
-from fink_broker.sparkUtils import connect_to_raw_database
+from fink_broker.spark_utils import init_sparksession
+from fink_broker.spark_utils import connect_to_raw_database
 from fink_broker.logging_utils import get_fink_logger, inspect_application
 from fink_broker.partitioning import convert_to_datetime
