+++ bin/save_distribution_schema.py
@@ -22,7 +22,7 @@ import glob
 import json
 
 from fink_broker.parser import getargs
-from fink_broker.sparkUtils import init_sparksession, load_parquet_files
+from fink_broker.spark_utils import init_sparksession, load_parquet_files
 from fink_broker.avro_utils import readschemafromavrofile
 from fink_broker.logging_utils import get_fink_logger, inspect_application
