+++ fink_broker/partitioning.py
@@ -43,7 +43,7 @@ def convert_to_millitime(jd: pd.Series, format=None, now=None):
 
     Examples
     --------
-    >>> from fink_broker.sparkUtils import load_parquet_files
+    >>> from fink_broker.spark_utils import load_parquet_files
     >>> df = load_parquet_files("online/raw")
     >>> df = df.withColumn('millis', convert_to_millitime(df['candidate.jd']))
     >>> pdf = df.select('millis').toPandas()
@@ -98,7 +98,7 @@ def convert_to_datetime(jd: pd.Series, format=None) -> pd.Series:
 
     Examples
     --------
-    >>> from fink_broker.sparkUtils import load_parquet_files
+    >>> from fink_broker.spark_utils import load_parquet_files
     >>> df = load_parquet_files("online/raw")
     >>> df = df.withColumn('datetime', convert_to_datetime(df['candidate.jd']))
     >>> pdf = df.select('datetime').toPandas()
@@ -129,7 +129,7 @@ def compute_num_part(df, partition_size=128.0):
 
     Examples
     --------
-    >>> from fink_broker.sparkUtils import load_parquet_files
+    >>> from fink_broker.spark_utils import load_parquet_files
     >>> df = load_parquet_files("online/raw")
     >>> numpart = compute_num_part(df, partition_size=128.)
     >>> print(numpart)
