+++ fink_broker/science.py
@@ -94,7 +94,7 @@ def ang2pix(ra: pd.Series, dec: pd.Series, nside: pd.Series) -> pd.Series:
 
     Examples
     --------
-    >>> from fink_broker.sparkUtils import load_parquet_files
+    >>> from fink_broker.spark_utils import load_parquet_files
     >>> df = load_parquet_files(ztf_alert_sample)
 
     >>> df_index = df.withColumn(
@@ -133,7 +133,7 @@ def ang2pix_array(ra: pd.Series, dec: pd.Series, nside: pd.Series) -> pd.Series:
 
     Examples
     --------
-    >>> from fink_broker.sparkUtils import load_parquet_files
+    >>> from fink_broker.spark_utils import load_parquet_files
     >>> df = load_parquet_files(ztf_alert_sample)
 
     >>> nsides = F.array([F.lit(256), F.lit(4096), F.lit(131072)])
@@ -201,7 +201,7 @@ def apply_science_modules(df: DataFrame, noscience: bool = False) -> DataFrame:
 
     Examples
     --------
-    >>> from fink_broker.sparkUtils import load_parquet_files
+    >>> from fink_broker.spark_utils import load_parquet_files
     >>> from fink_broker.logging_utils import get_fink_logger
     >>> logger = get_fink_logger('raw2cience_test', 'INFO')
     >>> _LOG = logging.getLogger(__name__)
@@ -474,7 +474,7 @@ def apply_science_modules_elasticc(df: DataFrame) -> DataFrame:
 
     Examples
     --------
-    >>> from fink_broker.sparkUtils import load_parquet_files
+    >>> from fink_broker.spark_utils import load_parquet_files
     >>> from fink_broker.logging_utils import get_fink_logger
     >>> logger = get_fink_logger('raw2cience_elasticc_test', 'INFO')
     >>> _LOG = logging.getLogger(__name__)
